+++ CHANGELOG.md
@@ -19,8 +19,14 @@ The **Unreleased** section collects notes for unreleased changes and features, u
 
 ## Unreleased
 
+
+## [7.3.192] - 2021-01-18
+
 - [Added] server information in admin panel, and moved server restart to it
 - [Added] after login or navigate to '/' redirect to "logo url" as home page
+
+## [7.3.191] - 2021-01-18
+
 
 ## [7.3.191] - 2021-01-18